backend/requirements-minimal.txt
# Minimal requirements for testing deployment
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
google-cloud-firestore==2.13.1
google-auth==2.25.2
httpx==0.25.2
python-dotenv==1.0.0
